+++ frontend/app/src/pages/people/index.tsx
