+++ frontend/app/src/pages/people/tabs/index.tsx
